feat(blog section): word wrap, styles and props
Adding a blog post with a very long title would cause the card layout
to increase it's size, making it inconsistent with the rest of the
cards. So I added a new CSS class and new tailwind classes to wrap the
title and also long tags. It was also added a `title` attribute so when
hovering with a mouse it's possible to read the longer text version.

Since the `BlogPosts` component is going to be an optional feature, I
added a new boolean prop called `enabled`, if this is false, the
component won't be rendered.

Some new placeholder image URLs and texts were introduced with this
patch, they are there to help better visualize the possibilities
within the component.