refactor(blog section): placeholder content
The images used as placeholder were removed, they are going to be
replaced by direct links.

The `enabled` property was set to false, as a default behavior the blog
posts won't be shown automatically with this patch. The title and tags
received new generic values to better express their uses.