fix: button props and EOL format
The props that the Button component receives are not optional and cannot
be null or undefined, with this patch I added nullish coalescing
operator and also a ternary to avoid the component crashing if it does
not receive a style or an icon. Affected file:

- `Button.jsx` (src/components)

My previous commits changed some files EOL format from CRLF to LF.
This patch is aimed to try to fix that. Affected files:

- `index.css` (src)
- `App.jsx` (src)
- `index.js` (src/constants)

Co-authored-by: Parth Mittal <76661350+mittal-parth@users.noreply.github.com>