OpenSource: Move fetchContributions to a Netlify function (#87)
When using environment variables via Vite, these tokens are exposed in the bundle output that is created for a production build upon running npm run build. This is not good as the token is exposed and anyone can get access to this token by inspecting the website.

Since we don't have a backend to store the env variables, added a Netlify function to fetch the contributions using the GitHub GraphQL API. This keeps the GitHub token (and others in future), outside the build output.

Updated the readme as well.