UI: Swap Achievements and Projects layout
I feel it's important to show the achievements more than the projects.

Currently the achievements section is a carousel of cards. This means the user has to put additional effort to press those buttons and look at the achievements.

Projects on the other hand are scrollable, meaning they are more likely to go through them.

The idea is to swap the layout where the actual card layout remains the same.