UI: Swap Achievements and Projects layout (#112)
I feel it's important to show the achievements more than the projects.

Earlier, the achievements section was a carousel of cards. This means the user has to put additional effort to press those buttons and look at the achievements.

Projects on the other hand were a scroll-able grid, meaning they are more likely to go through them.

This commit makes the achievements section a scroll-able grid and the projects section a carousel.